Remove logging crate
