+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "frida-app",
     "crates/benchmark",
-    "logging",
     "crates/common",
     "defrida-app",
 ]
@@ -22,7 +21,6 @@ ed25519-dalek = { version = "2.0.0", features = ["rand_core"] }
 rand_core = { version = "0.6.4", features = ["getrandom"] }
 frida-app = { path = "./frida-app" }
 common = { path = "./crates/common" }
-logging = { path = "./logging" }
 tracing = "0.1.41"
 tracing-subscriber = "0.3.19"
 tracing-appender = "0.2.3"
