--- logging/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "logging"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-tracing-appender = { workspace = true }
